Add update user functions

+++ src/main/kotlin/uk/co/culturebook/modules/culture/data/database/repositories/ContributionRepository.kt
@@ -102,6 +102,15 @@ object ContributionRepository : ContributionDao {
         bearer: String,
         fileHost: String
     ): Boolean {
+        val emptyResponse = client.post(MediaRoute.BucketRoute.emptyBucket(fileHost, request.id)) {
+            headers {
+                append(Constants.Headers.Authorization, "Bearer $bearer")
+                append(Constants.Headers.ApiKey, apiKey)
+            }
+            contentType(ContentType.Application.Json)
+            setBody(request)
+        }.status == HttpStatusCode.OK
+        if (!emptyResponse) return false
         val response = client.delete(MediaRoute.BucketRoute.getBucket(fileHost, request.id)) {
             headers {
                 append(Constants.Headers.Authorization, "Bearer $bearer")
@@ -178,7 +187,8 @@ object ContributionRepository : ContributionDao {
                 { FavouriteContributions.userId eq userId }
             )
             .select {
-                val similarity = if (searchString.isNotBlank()) Similarity(Contributions.name, searchString) greaterEq 0.25 else null
+                val similarity =
+                    if (searchString.isNotBlank()) Similarity(Contributions.name, searchString) greaterEq 0.25 else null
                 val blocked = BlockedContributions.id.isNull()
                 val inType = Contributions.type inList types.map { it.toString() }
                 val inElement = Contributions.element_id eq elementId
